подготовил сервис для тестового запуска (без модуля psql)

+++ internal/services/aggMileageHours/transport_models.go
@@ -2,6 +2,6 @@ package aggmileagehours
 
 // события для отправки в горутину агрегации
 type eventForAgg struct {
-	offset    int
+	offset    int64
 	eventData *eventData
 }
